Breakout network utils into zeroconf.utils.net
